- reorganize commands #53
